+++ src/models/webhook_events.rs
@@ -981,37 +981,96 @@ pub struct InstallationEventRepository {
 
 #[cfg(test)]
 mod tests {
+    use chrono::TimeZone;
+
+    use crate::models::AuthorAssociation;
+
     use super::payload::*;
     use super::*;
 
     #[test]
+    fn deserialize_commit_comment_created() {
+        let json = include_str!("../../tests/resources/commit_comment_created_webhook_event.json");
+        let event = WebhookEvent::try_from_header_and_body("commit_comment", json).unwrap();
+        let WebhookEventPayload::CommitComment(commit_comment_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
+        assert_eq!(
+            commit_comment_event.action,
+            CommitCommentWebhookEventAction::Created
+        );
+        assert_eq!(
+            commit_comment_event.comment.author_association,
+            AuthorAssociation::Owner
+        );
+        assert_eq!(
+            commit_comment_event.comment.body.as_deref(),
+            Some("@gagbo-test-app[bot] compare-tag v0.1")
+        );
+    }
+
+    #[test]
     fn deserialize_installation_created() {
         let json = include_str!("../../tests/resources/installation_created_webhook_event.json");
-        let event = WebhookEventType::Installation
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::Installation(install_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("installation", json).unwrap();
+        let WebhookEventPayload::Installation(install_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             install_event.action,
             InstallationWebhookEventAction::Created
         );
-        assert_eq!(install_event.repositories.len(), 3);
-        assert_eq!(install_event.repositories[2].name, "octocrab");
+        assert_eq!(install_event.repositories.as_ref().unwrap().len(), 3);
+        assert_eq!(install_event.repositories.unwrap()[2].name, "octocrab");
     }
 
     #[test]
     fn deserialize_installation_deleted() {
         let json = include_str!("../../tests/resources/installation_deleted_webhook_event.json");
-        let event = WebhookEventType::Installation
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::Installation(install_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("installation", json).unwrap();
+        let WebhookEventPayload::Installation(install_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             install_event.action,
             InstallationWebhookEventAction::Deleted
         );
-        assert_eq!(install_event.repositories.len(), 3);
-        assert_eq!(install_event.repositories[2].name, "octocrab");
+        assert_eq!(install_event.repositories.as_ref().unwrap().len(), 3);
+        assert_eq!(install_event.repositories.unwrap()[2].name, "octocrab");
+    }
+
+    #[test]
+    fn deserialize_installation_new_permissions_accepted() {
+        let json = include_str!(
+            "../../tests/resources/installation_new_permissions_accepted_webhook_event.json"
+        );
+        let event = WebhookEvent::try_from_header_and_body("installation", json).unwrap();
+        let WebhookEventPayload::Installation(ref install_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
+        let Some(EventInstallation::Full(installation)) = event.installation else {
+            panic!(
+                "event is missing a fully described installation object {:?}",
+                event
+            )
+        };
+        assert_eq!(
+            install_event.action,
+            InstallationWebhookEventAction::NewPermissionsAccepted
+        );
+        assert_eq!(
+            installation.updated_at.unwrap(),
+            chrono::Utc
+                .with_ymd_and_hms(2023, 8, 18, 13, 28, 4)
+                .unwrap()
+        );
+        assert_eq!(
+            installation.created_at.unwrap(),
+            chrono::Utc
+                .with_ymd_and_hms(2023, 7, 13, 9, 35, 31)
+                .unwrap()
+        );
+        assert_eq!(installation.events.len(), 12);
     }
 
     #[test]
@@ -1019,10 +1078,13 @@ mod tests {
         let json = include_str!(
             "../../tests/resources/installation_repositories_removed_webhook_event.json"
         );
-        let event = WebhookEventType::InstallationRepositories
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::InstallationRepositories(install_repositories_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event =
+            WebhookEvent::try_from_header_and_body("installation_repositories", json).unwrap();
+        let WebhookEventPayload::InstallationRepositories(install_repositories_event) =
+            event.specific
+        else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             install_repositories_event.action,
             InstallationRepositoriesWebhookEventAction::Removed
@@ -1037,10 +1099,10 @@ mod tests {
     #[test]
     fn deserialize_issue_comment_created() {
         let json = include_str!("../../tests/resources/issue_comment_created_webhook_event.json");
-        let event = WebhookEventType::IssueComment
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::IssueComment(issue_comment_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("issue_comment", json).unwrap();
+        let WebhookEventPayload::IssueComment(issue_comment_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             issue_comment_event.action,
             IssueCommentWebhookEventAction::Created
@@ -1051,10 +1113,10 @@ mod tests {
     #[test]
     fn deserialize_issue_comment_deleted() {
         let json = include_str!("../../tests/resources/issue_comment_deleted_webhook_event.json");
-        let event = WebhookEventType::IssueComment
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::IssueComment(issue_comment_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("issue_comment", json).unwrap();
+        let WebhookEventPayload::IssueComment(issue_comment_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             issue_comment_event.action,
             IssueCommentWebhookEventAction::Deleted
@@ -1064,10 +1126,10 @@ mod tests {
     #[test]
     fn deserialize_issue_comment_edited() {
         let json = include_str!("../../tests/resources/issue_comment_edited_webhook_event.json");
-        let event = WebhookEventType::IssueComment
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::IssueComment(issue_comment_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("issue_comment", json).unwrap();
+        let WebhookEventPayload::IssueComment(issue_comment_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             issue_comment_event.action,
             IssueCommentWebhookEventAction::Edited
@@ -1078,40 +1140,40 @@ mod tests {
     #[test]
     fn deserialize_issues_labeled() {
         let json = include_str!("../../tests/resources/issues_labeled_webhook_event.json");
-        let event = WebhookEventType::Issues
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::Issues(issues_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("issues", json).unwrap();
+        let WebhookEventPayload::Issues(issues_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(issues_event.action, IssuesWebhookEventAction::Labeled);
     }
 
     #[test]
     fn deserialize_issues_opened() {
         let json = include_str!("../../tests/resources/issues_opened_webhook_event.json");
-        let event = WebhookEventType::Issues
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::Issues(issues_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("issues", json).unwrap();
+        let WebhookEventPayload::Issues(issues_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(issues_event.action, IssuesWebhookEventAction::Opened);
     }
 
     #[test]
     fn deserialize_ping() {
         let json = include_str!("../../tests/resources/ping_webhook_event.json");
-        let event = WebhookEventType::Ping
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::Ping(ping_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("ping", json).unwrap();
+        let WebhookEventPayload::Ping(ping_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(ping_event.hook.unwrap().id, 423885699);
     }
 
     #[test]
     fn deserialize_pull_request_closed() {
         let json = include_str!("../../tests/resources/pull_request_closed_webhook_event.json");
-        let event = WebhookEventType::PullRequest
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::PullRequest(pull_request_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("pull_request", json).unwrap();
+        let WebhookEventPayload::PullRequest(pull_request_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             pull_request_event.action,
             PullRequestWebhookEventAction::Closed
@@ -1122,10 +1184,10 @@ mod tests {
     #[test]
     fn deserialize_pull_request_opened() {
         let json = include_str!("../../tests/resources/pull_request_opened_webhook_event.json");
-        let event = WebhookEventType::PullRequest
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::PullRequest(pull_request_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("pull_request", json).unwrap();
+        let WebhookEventPayload::PullRequest(pull_request_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             pull_request_event.action,
             PullRequestWebhookEventAction::Opened
@@ -1136,10 +1198,10 @@ mod tests {
     fn deserialize_pull_request_synchronize() {
         let json =
             include_str!("../../tests/resources/pull_request_synchronize_webhook_event.json");
-        let event = WebhookEventType::PullRequest
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::PullRequest(pull_request_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("pull_request", json).unwrap();
+        let WebhookEventPayload::PullRequest(pull_request_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             pull_request_event.action,
             PullRequestWebhookEventAction::Synchronize
@@ -1149,13 +1211,23 @@ mod tests {
     #[test]
     fn deserialize_repository_deleted() {
         let json = include_str!("../../tests/resources/repository_deleted_webhook_event.json");
-        let event = WebhookEventType::Repository
-            .parse_specific_payload(serde_json::from_str(json).unwrap())
-            .unwrap();
-        let WebhookEventPayload::Repository(repository_event) = event else {panic!(" event is of the wrong type {:?}", event)};
+        let event = WebhookEvent::try_from_header_and_body("repository", json).unwrap();
+        let WebhookEventPayload::Repository(repository_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
         assert_eq!(
             repository_event.action,
             RepositoryWebhookEventAction::Deleted
         );
+    }
+
+    #[test]
+    fn deserialize_push() {
+        let json = include_str!("../../tests/resources/push_webhook_event.json");
+        let event = WebhookEvent::try_from_header_and_body("push", json).unwrap();
+        let WebhookEventPayload::Push(push_event) = event.specific else {
+            panic!(" event is of the wrong type {:?}", event)
+        };
+        assert!(push_event.created);
     }
 }